[SkPath] Increase GenerationID to 64 bits
* GenerationID was originally intended to be a modification tracker per object but became conflated with global uniqueness. This was fine so long as ID usages were sufficiently temporally separated that wrapping IDs would never collide. However, at 32 bits it is feasible for an attacker to cause wrapping in a relatively small amount of time.

* This patch ups the GenerationID to 64 bits

Bug: b/562038679
Bug: b/556268833
Reviewed-on: https://skia-review.googlesource.com/c/skia/+/1366216
Commit-Queue: Thomas Smith <thomsmit@google.com>

+++ include/core/SkPath.h
@@ -1088,7 +1088,7 @@ public:
 
         example: https://fiddle.skia.org/c/@Path_getGenerationID
     */
-    uint32_t getGenerationID() const;
+    uint64_t getGenerationID() const;
 
     /** Returns if SkPath data is consistent. Corrupt SkPath data is detected if
         internal values are out of range or internal storage does not match
